+++ docs/CHANGELOG.rst
@@ -13,6 +13,7 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 Changed
 -------
 - Change fallback for storage resource from 200GB to 10GB
+- Restrict processing container from using service account permissions
 
 
 ===================